Added more friendships to the network and fixed some wrong code

+++ tuple.c
@@ -1,6 +1,6 @@
 #include "tuple.h"
 
-void Set(Tuple* t, int a, int b, int c) {
+void Set(Tuple *t, int a, int b, double c) {
     t->a = a;
     t->b = b;
     t->c = c;
@@ -14,11 +14,11 @@ int Second(Tuple t) {
     return t.b;
 }
 
-int Third(Tuple t) {
+double Third(Tuple t) {
     return t.c;
 }
 
-int compTuple(const void* a,const void* b) {
+int compTuple(const void *a,const void *b) {
 
     Tuple c = *((Tuple*) a);
     Tuple d = *((Tuple*) b);
@@ -26,5 +26,4 @@ int compTuple(const void* a,const void* b) {
     if (Third(c) < Third(d)) return -1;
     if (Third(c) > Third(d)) return 1;
     else return 0;
-
 }
